+++ contracts/multisig-prover/src/encoding/mod.rs
@@ -5,8 +5,8 @@ mod stellar_xdr;
 use axelar_wasm_std::hash::Hash;
 use cosmwasm_std::HexBinary;
 use error_stack::Result;
-use multisig::msg::SignerWithSig;
-use multisig::verifier_set::VerifierSet;
+use multisig::SignerWithSig;
+use multisig::VerifierSet;
 use multisig_prover_api::encoding::Encoder;
 
 use crate::error::ContractError;
