+++ contracts/multisig/src/exported.rs
@@ -1,0 +1,8 @@
+pub use crate::client::Client;
+pub use crate::key::KeyType::Ecdsa;
+pub use crate::key::{KeyType,KeyTyped,Signature,PublicKey};
+pub use crate::msg::Signer;
+pub use crate::msg::{ExecuteMsg,QueryMsg,InstantiateMsg, SignerWithSig};
+pub use crate::multisig::Multisig;
+pub use crate::types::MultisigState;
+pub use crate::verifier_set::VerifierSet;
