+++ contracts/multisig/src/lib.rs
@@ -1,15 +1,16 @@
 mod client;
-pub use client::Client;
 pub mod contract;
-pub mod error;
+mod error;
 pub mod events;
 pub mod key;
 pub mod msg;
-pub mod multisig;
+mod multisig;
 mod signing;
 mod state;
 pub mod types;
 pub mod verifier_set;
+mod exported;
+pub use exported::*;
 
 #[cfg(feature = "secp256k1")]
 mod secp256k1;
@@ -20,4 +21,4 @@ mod ed25519;
 #[cfg(any(test, feature = "test"))]
 pub mod test;
 
-pub use crate::error::ContractError;
+use crate::error::ContractError;
