+++ contracts/voting-verifier/src/events.rs
@@ -346,7 +346,7 @@ mod test {
     use axelar_wasm_std::{nonempty, Threshold, VerificationStatus};
     use cosmwasm_std::testing::MockApi;
     use cosmwasm_std::{Attribute, Uint128};
-    use multisig::key::KeyType;
+    use multisig::KeyType;
     use multisig::test::common::{build_verifier_set, ecdsa_test_data};
     use multisig::verifier_set::VerifierSet;
     use router_api::{CrossChainId, Message};
